wlr_cursor

+++ examples/output-layout.c
@@ -22,6 +22,7 @@
 #include <wlr/types/wlr_keyboard.h>
 #include <math.h>
 #include "shared.h"
+#include "config.h"
 #include "cat.h"
 
 struct sample_state {
@@ -49,8 +50,8 @@ static void handle_output_frame(struct output_state *output, struct timespec *ts
 		float matrix[16];
 
 		// transform global coordinates to local coordinates
-		int local_x = sample->x_offs;
-		int local_y = sample->y_offs;
+		double local_x = sample->x_offs;
+		double local_y = sample->y_offs;
 		wlr_output_layout_output_coords(sample->layout, output->output, &local_x,
 				&local_y);
